// web/src/api/openapi-schema/adminOCRAssetListStatus.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * storyden
 * Storyden social API for building community driven platforms.
The Storyden API does not adhere to semantic versioning but instead applies a rolling strategy with deprecations and minimal breaking changes. This has been done mainly for a simpler development process and it may be changed to a more fixed versioning strategy in the future. Ultimately, the primary way Storyden tracks versions is dates, there are no set release tags currently.

 * OpenAPI spec version: v1.26.13-post
 */

export type AdminOCRAssetListStatus =
  (typeof AdminOCRAssetListStatus)[keyof typeof AdminOCRAssetListStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AdminOCRAssetListStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
  skipped: "skipped",
} as const;